# backend/openapi.yml
openapi: 3.0.3
info:
  title: CrowdPulse - OpenAPI 3.0
  description: "CrowdPulse API specifics"
  version: 1.0.0
servers:
  - url: http://localhost:4000
...
paths:
...
  /v1/sentiment/timeline:
    get:
      tags: [v1]
      summary: Gets the sentiment timeline values.
      parameters:
        - name: dbs
          in: query
          description: MongoDB database names
          required: true
          schema:
            type: array
        - name: algorithm
          in: query
          description: sent-it, feel-it or hate-speech
          required: true
          schema:
            type: string
        - name: dateFrom
          in: query
          description: ISO date like 2022-01-09T00:00:00.000Z (dateTo required)
          required: false
          schema:
            type: string
        - name: dateTo
          in: query
          description: ISO date like 2022-01-09T00:00:00.000Z (dateFrom required)
          required: false
          schema:
            type: string
        - name: tags
          in: query
          description: A list of tags
          required: false
          schema:
            type: array
        - name: processedText
          in: query
          description: A list processed words
          required: false
          schema:
            type: array
        - name: hashtags
          in: query
          description: A list of hashtags (# not required)
          required: false
          schema:
            type: array
        - name: usernames
          in: query
          description: A list of usernames
          required: false
          schema:
            type: array
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1_sentiment_timeline'
        '400':
          description: missing query parameter
        '500':
          description: Executing error
  # [v1] GET - /v1/word
...
components:
  schemas:
...
    v1_sentiment_timeline:
      type: object
      properties:
        algorithm:
          type: string
        data:
          type: object
          example: [
            {
              "sentimentPositiveCount": number,
              "sentimentNeutralCount": number,
              "sentimentNegativeCount": number,
              "emotionJoyCount": number,
              "emotionSadnessCount": number,
              "emotionAngerCount": number,
              "emotionFearCount": number,
              "hateSpeechAcceptableCount": number,
              "hateSpeechInappropriateCount": number,
              "hateSpeechOffensiveCount": number,
              "hateSpeechViolentCount": number,
              "date": '2023-02-21'
            }
          ]
...